style on edit inside the singlePost and movile responsive
