+++ client/src/components/Blog/SinglePost.jsx
@@ -5,6 +5,13 @@ import axios from 'axios';
 import 'react-quill/dist/quill.snow.css';
 import { useReactToPrint } from 'react-to-print';
 import EditPostCard from './EditPostCard';
+import {
+    FacebookShareButton,
+    TwitterShareButton,
+    LinkedinShareButton,
+  } from 'react-share';
+  import { FaFacebook, FaTwitter, FaLink, FaLinkedin } from 'react-icons/fa';
+  
 
 export const SinglePost = ({ onEdit }) => {
   const [post, setPost] = useState(null);
@@ -12,6 +19,12 @@ export const SinglePost = ({ onEdit }) => {
   const [isEditing, setIsEditing] = useState(false);
   const { id } = useParams();
   const { isLoggedIn, user } = useContext(AuthContext);
+
+  const [isCopied, setIsCopied] = useState(false);
+  const copyToClipboard = () => {
+    navigator.clipboard.writeText(window.location.href);
+    setIsCopied(true);
+  };
 
   useEffect(() => {
     axios.get(`/api/post/${id}`).then((response) => {
@@ -42,7 +55,7 @@ export const SinglePost = ({ onEdit }) => {
     <>
       <div
         ref={componentRef}
-        className="flex justify-center items-center relative bg-gray-200 pt-5 pb-10"
+        className="flex justify-center items-center relative bg-gray-200 pt-5 pb-10 font-pop"
       >
         {postBeingEdited && (
           <div className="max-w-2xl mt-5 bg-gray-50 rounded-2xl overflow-hidden shadow-lg ">
@@ -77,12 +90,34 @@ export const SinglePost = ({ onEdit }) => {
             >
               Print
             </span>
+            <div className="my-1 space-x-4">
+            <FacebookShareButton
+              quote={post?.title}
+              url={window.location.href}
+              hashtag="#myblog"
+            >
+              <FaFacebook size={20} />
+            </FacebookShareButton>
+            <TwitterShareButton
+              title={post?.title}
+              url={window.location.href}
+              via="@myblog"
+            >
+              <FaTwitter size={20} />
+            </TwitterShareButton>
+            <LinkedinShareButton>
+              <FaLinkedin size={20} />
+            </LinkedinShareButton>
+            <button onClick={copyToClipboard}>
+              {isCopied ? 'Copied!' : <FaLink />}
+            </button>
+          </div>
             {isLoggedIn && user?._id === post?.author?._id && (
               <button
                 onClick={handleEdit}
-                className="bg-cyan-500 text-white font-medium px-8 py-1 my-1 rounded-full mt-2"
+                className="bg-cyan-500 text-white font-medium px-8 py-1 my-1 rounded mt-2"
               >
-                Edit
+               {!isEditing ? "Edit" : "finish edition"} 
               </button>
             )}
             {isEditing && (
